+++ jetcd-core/src/test/java/io/etcd/jetcd/internal/impl/AuthClientTest.java
@@ -27,8 +27,8 @@ import io.etcd.jetcd.auth.AuthRoleListResponse;
 import io.etcd.jetcd.auth.Permission;
 import io.etcd.jetcd.auth.Permission.Type;
 import io.etcd.jetcd.data.ByteSequence;
-import io.etcd.jetcd.internal.infrastructure.EtcdCluster;
-import io.etcd.jetcd.internal.infrastructure.EtcdClusterFactory;
+import io.etcd.jetcd.launcher.EtcdCluster;
+import io.etcd.jetcd.launcher.EtcdClusterFactory;
 import java.io.IOException;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
