+++ jetcd-core/src/test/java/io/etcd/jetcd/internal/impl/ClientConnectionManagerTest.java
@@ -22,8 +22,8 @@ import com.google.common.base.Charsets;
 import io.etcd.jetcd.Client;
 import io.etcd.jetcd.ClientBuilder;
 import io.etcd.jetcd.data.ByteSequence;
-import io.etcd.jetcd.internal.infrastructure.EtcdClusterResource;
 import io.etcd.jetcd.kv.PutResponse;
+import io.etcd.jetcd.launcher.junit.EtcdClusterResource;
 import io.grpc.CallOptions;
 import io.grpc.Channel;
 import io.grpc.ClientCall;
