+++ jetcd-core/src/test/java/io/etcd/jetcd/internal/impl/LoadBalancerTest.java
@@ -20,8 +20,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.etcd.jetcd.Client;
 import io.etcd.jetcd.KV;
 import io.etcd.jetcd.data.Response;
-import io.etcd.jetcd.internal.infrastructure.EtcdClusterResource;
 import io.etcd.jetcd.kv.PutResponse;
+import io.etcd.jetcd.launcher.junit.EtcdClusterResource;
 import io.grpc.PickFirstBalancerFactory;
 import io.grpc.util.RoundRobinLoadBalancerFactory;
 import java.util.List;
