+++ jetcd-core/src/test/java/io/etcd/jetcd/internal/impl/SslTest.java
@@ -21,7 +21,8 @@ import com.google.common.base.Charsets;
 import io.etcd.jetcd.Client;
 import io.etcd.jetcd.KV;
 import io.etcd.jetcd.data.ByteSequence;
-import io.etcd.jetcd.internal.infrastructure.EtcdClusterResource;
+import io.etcd.jetcd.launcher.junit.EtcdClusterResource;
+import io.etcd.jetcd.launcher.TestConstants;
 import io.grpc.netty.GrpcSslContexts;
 import java.io.File;
 import java.io.FileInputStream;
