+++ jetcd-core/src/test/java/io/etcd/jetcd/internal/impl/TestUtil.java
@@ -19,10 +19,6 @@ import com.google.common.base.Charsets;
 import io.etcd.jetcd.data.ByteSequence;
 import java.io.IOException;
 import java.net.ServerSocket;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-import org.testcontainers.containers.GenericContainer;
 
 public class TestUtil {
 
@@ -38,31 +34,5 @@ public class TestUtil {
     try (ServerSocket socket = new ServerSocket(0)) {
       return socket.getLocalPort();
     }
-  }
-
-  public static String buildClientEndpoint(GenericContainer container) {
-    return buildEndpoint(container, "http", TestConstants.ETCD_CLIENT_PORT);
-  }
-
-  public static List<String> buildClientEndpoints(GenericContainer... etcdContainers) {
-    return Arrays.stream(etcdContainers)
-            .map(TestUtil::buildClientEndpoint)
-            .collect(Collectors.toList());
-  }
-
-  public static String buildPeerEndpoint(GenericContainer container) {
-    return buildEndpoint(container, "http", TestConstants.ETCD_PEER_PORT);
-  }
-
-  public static List<String> buildPeerEndpoints(GenericContainer... etcdContainers) {
-    return Arrays.stream(etcdContainers)
-            .map(TestUtil::buildPeerEndpoint)
-            .collect(Collectors.toList());
-  }
-
-  public static String buildEndpoint(GenericContainer container, String scheme, int port) {
-    String nodeAddress = container.getContainerIpAddress();
-    Integer nodePort = container.getMappedPort(port);
-    return scheme + "://" + nodeAddress + ":" + nodePort;
   }
 }
